# packages/schemas/openapi.json
{
  "components": {
    "schemas": {
      "AcceptedEventResponse": {
        "properties": {
          "actor_player_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Player Id"
          },
          "created_at": {
            "title": "Created At"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "game_id": {
            "format": "uuid",
            "title": "Game Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "state_hash": {
            "title": "State Hash",
            "type": "string"
          }
        },
        "required": [
          "id",
          "game_id",
          "sequence",
          "actor_player_id",
          "event_type",
          "payload",
          "state_hash",
          "created_at"
        ],
        "title": "AcceptedEventResponse",
        "type": "object"
      },
      "ActionAcceptedResponse": {
        "properties": {
          "accepted_events": {
            "items": {
              "$ref": "#/components/schemas/AcceptedEventResponse"
            },
            "title": "Accepted Events",
            "type": "array"
          },
          "event_sequence": {
            "title": "Event Sequence",
            "type": "integer"
          },
          "game_id": {
            "format": "uuid",
            "title": "Game Id",
            "type": "string"
          },
          "state": {
            "additionalProperties": true,
            "title": "State",
            "type": "object"
          },
          "state_hash": {
            "title": "State Hash",
            "type": "string"
          },
          "status": {
            "const": "accepted",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "game_id",
          "accepted_events",
          "state",
          "state_hash",
          "event_sequence"
        ],
        "title": "ActionAcceptedResponse",
        "type": "object"
      },
...
      "CreateGameRequest": {
        "additionalProperties": false,
        "properties": {
          "players": {
            "items": {
              "$ref": "#/components/schemas/PlayerCreateRequest"
            },
            "maxItems": 5,
            "minItems": 2,
            "title": "Players",
            "type": "array"
          },
          "seed": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "settings": {
            "additionalProperties": true,
            "title": "Settings",
            "type": "object"
          }
        },
        "required": [
          "players"
        ],
        "title": "CreateGameRequest",
        "type": "object"
      },
...
      "GameMetadataResponse": {
        "properties": {
          "created_at": {
            "title": "Created At"
          },
          "current_phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Phase"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/PlayerRecordResponse"
            },
            "title": "Players",
            "type": "array"
          },
          "ruleset_version": {
            "title": "Ruleset Version",
            "type": "string"
          },
          "seed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "settings": {
            "additionalProperties": true,
            "title": "Settings",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "status",
          "ruleset_version",
          "seed",
          "current_phase",
          "settings",
          "players",
          "created_at",
          "updated_at"
        ],
        "title": "GameMetadataResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LegalActionResponse": {
        "properties": {
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "expected_event_sequence": {
            "title": "Expected Event Sequence",
            "type": "integer"
          },
          "expected_state_hash": {
            "title": "Expected State Hash",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "schema": {
            "additionalProperties": true,
            "title": "Schema",
            "type": "object"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "type",
          "payload",
          "expected_state_hash",
          "expected_event_sequence",
          "schema"
        ],
        "title": "LegalActionResponse",
        "type": "object"
      },
      "LegalActionsResponse": {
        "properties": {
          "actor_player_id": {
            "format": "uuid",
            "title": "Actor Player Id",
            "type": "string"
          },
          "event_sequence": {
            "title": "Event Sequence",
            "type": "integer"
          },
          "game_id": {
            "format": "uuid",
            "title": "Game Id",
            "type": "string"
          },
          "legal_actions": {
            "items": {
              "$ref": "#/components/schemas/LegalActionResponse"
            },
            "title": "Legal Actions",
            "type": "array"
          },
          "state_hash": {
            "title": "State Hash",
            "type": "string"
          }
        },
        "required": [
          "game_id",
          "actor_player_id",
          "legal_actions",
          "state_hash",
          "event_sequence"
        ],
        "title": "LegalActionsResponse",
        "type": "object"
      },
...
      "PlayerCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "human",
              "ai"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "kind"
        ],
        "title": "PlayerCreateRequest",
        "type": "object"
      },
      "PlayerRecordResponse": {
        "properties": {
          "controller_type": {
            "title": "Controller Type",
            "type": "string"
          },
          "created_at": {
            "title": "Created At"
          },
          "game_id": {
            "format": "uuid",
            "title": "Game Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "seat_order": {
            "title": "Seat Order",
            "type": "integer"
          },
          "state": {
            "additionalProperties": true,
            "title": "State",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "game_id",
          "seat_order",
          "name",
          "controller_type",
          "status",
          "state",
          "created_at",
          "updated_at"
        ],
        "title": "PlayerRecordResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Local FastAPI foundation for the Monopoly AI research game.",
    "title": "Monopoly AI Game API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/games": {
      "post": {
        "operationId": "create_game_games_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameMetadataResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Game",
        "tags": [
          "games"
        ]
      }
    },
...
    "/games/{game_id}/actions": {
      "post": {
        "operationId": "submit_action_games__game_id__actions_post",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ActionAcceptedResponse"
                    },
                    {
                      "additionalProperties": true,
                      "type": "object"
                    }
                  ],
                  "title": "Response Submit Action Games  Game Id  Actions Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Action",
        "tags": [
          "games"
        ]
      }
    },
...
    "/games/{game_id}/legal-actions": {
      "get": {
        "operationId": "get_legal_actions_games__game_id__legal_actions_get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Game Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actor_player_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Actor Player Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegalActionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Legal Actions",
        "tags": [
          "games"
        ]
      }
    },
...
  }
}